huge refactoring for achievements & project structure
