+++ package.json
@@ -24,7 +24,7 @@
     "circular-buffer": "^1.0.3",
     "command-line-args": "^5.1.1",
     "convict": "^6.0.1",
-    "dateformat": "^3.0.3",
+    "dateformat": "^4.5.1",
     "discord": "^0.8.2",
     "discord-akairo": "^8.1.0",
     "discord.js": "^12.5.3",
@@ -34,7 +34,7 @@
     "memoizer-ts": "^1.0.1",
     "moment": "^2.29.1",
     "moment-timezone": "^0.5.33",
-    "node-schedule": "^1.3.2",
+    "node-schedule": "^2.0.0",
     "node-teamspeak-api": "^1.0.0",
     "nodemon": "^2.0.7",
     "singly-linked-list": "^1.4.0",
