+++ src/commands/achievements/ListAchievements.ts
@@ -2,27 +2,27 @@ import * as discord from "discord.js";
 import { BotgartCommand } from "../../BotgartCommand";
 
 /**
-Testcases:
+ Testcases:
 
-*/
+ */
 export class ListAchievements extends BotgartCommand {
-     constructor() {
+    constructor() {
         super("listachievements", {
-            aliases: ["lsachievements"],
-            quoted: true,
-            args: []
-        },
-        {
-          availableAsDM: true
-        }
+                aliases: ["listachievements", "lsachievements"],
+                quoted: true,
+                args: []
+            },
+            {
+                availableAsDM: true
+            }
         );
     }
 
     command(message: discord.Message, responsible: discord.User, guild: discord.Guild, args: any): void {
-        message.reply(this.getBotgartClient().getAchievements()
-                          .map(a => `\`${a.name}\`: ${a.getDescription()}`)
-                          .join("\n")
-                      , {"split": true});
+        message.reply(this.getBotgartClient().achievementRegistry.getAchievements()
+                .map(a => `\`${a.name}\`: ${a.getDescription()}`)
+                .join("\n")
+            , {"split": true});
     }
 }
