+++ src/listeners/IgnoringRolesListener.ts
@@ -21,7 +21,7 @@ export class IgnoringRolesListener extends Listener {
             const userdata = client.registrationRepository.getUserByDiscordId(newMember.user);
             let deletedLeads = 0;
             let revokedAchievements = 0;
-            for(const achievement of client.getAchievements()) {
+            for(const achievement of client.achievementRegistry.getAchievements()) {
                 const role: discord.Role | undefined = newMember.guild.roles.cache.find(r => r.name === achievement.getRoleName());
                 if(role === undefined) {
                     U.log("warning", `Could not find a role ${achievement.getRoleName()} on server ${newMember.guild.name} to remove when user ${newMember.displayName} chose to ignore achievement roles.`);
