+++ src/listeners/PermanentRoleListener.ts
@@ -17,7 +17,7 @@ export class PermanentRoleListener extends Listener {
             let role = g.roles.find(r => r.name === roleName);
             if(role) {
                 member.addRole(role, "permanent role");    
-                log("info", "Succesfully readded permanent role {0} to user {1}.".formatUnicorn(roleName, member.user.username));
+                log("info", "Succesfully re-added permanent role {0} to user {1}.".formatUnicorn(roleName, member.user.username));
             } else {
                 log("warning", "No role with name '{0}' was found on this server, which was configured as permanent role for user {1}. Skipping.".formatUnicorn(roleName, member.user.username));
             }
